Add ProjectFluent to manifest and move translation code to sub-namespace.
With ProjectFluent as a non-required dependency, it will be loaded before sd-access if available.
Moved translation code to new namespace and updated references for organization / consistency
