+++ stardew-access/Features/DynamicTiles.cs
@@ -3,6 +3,7 @@ using StardewValley;
 using StardewValley.Buildings;
 using StardewValley.Locations;
 using System.Text.Json;
+using stardew_access.Translation;
 using static stardew_access.Utils.JsonLoader;
 
 namespace stardew_access.Utils
