+++ stardew-access/Features/GameStateNarrator.cs
@@ -1,5 +1,6 @@
 ﻿using StardewValley;
 using System.Text.RegularExpressions;
+using stardew_access.Translation;
 
 namespace stardew_access.Utils
 {
