+++ stardew-access/Features/ReadTile.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 using StardewValley;
+using stardew_access.Translation;
 
 namespace stardew_access.Utils
 {
