+++ stardew-access/Patches/MenuWithInventoryPatches/ItemGrabMenuPatch.cs
@@ -1,3 +1,4 @@
+using stardew_access.Translation;
 using stardew_access.Utils;
 using StardewValley;
 using StardewValley.Menus;
