+++ stardew-access/Patches/MiniGamesPatches/IntroPatch.cs
@@ -1,4 +1,5 @@
 using StardewValley.Minigames;
+using stardew_access.Translation;
 
 namespace stardew_access.Patches
 {
