+++ stardew-access/Utils/InventoryUtils.cs
@@ -1,5 +1,6 @@
 using StardewValley;
 using StardewValley.Menus;
+using stardew_access.Translation;
 
 namespace stardew_access.Utils
 {
